Update feywild_rpg_console.py
finally. okay I've been messing around with this for a while and I finally got it to work. it was just outputting nothing at all so I had to do this with it. I tried a whole bunch of stuff. not done yet but this should include some more in-depth character generation.